Mostly Working - some small problems
